refactor: decouple selection logic by moving useBookSelection to hooks directory
- Relocate useBookSelection hook to src/features/Bookshelf/hooks/useBookSelection.ts.
- Keep BookSelectionProvider in src/features/Bookshelf/components/BookSelectionContext.tsx.
- Improve separation of concerns between UI providers and business logic hooks.
- Update all references in BookGrid, BookCard, and test files.
