+++ src/features/Bookshelf/components/BookGrid.tsx
@@ -21,10 +21,10 @@ import { Grid } from "react-window";
 import { useAppDispatch, useAppSelector } from "../../../store/store";
 import type { Book, BookWithState } from "../../../types/DatabaseModels";
 import { updateSettings } from "../../Settings/slice";
+import { useBookSelection } from "../hooks/useBookSelection";
 import { fetchBooksInSelectedBookshelf } from "../slice";
 import { andSearch, sortBy } from "../utils/BookshelfUtils";
 import BookCard from "./BookCard";
-import { useBookSelection } from "./BookSelectionContext";
 import AddBooksToBookshelvesDialog from "./Dialog/AddBooksToBookshelvesDialog";
 import BookDeleteDialog from "./Dialog/BookDeleteDialog";
 import SetBookTagsDialog from "./Dialog/SetBookTagsDialog";
